Managed to add feature where clicking on login button blurs current page
TODO: fix action when one clicks on the signup button when on the modal. currently, when one clicks one the signup button when on the modal, the signup page opens in background, but the modal stays open. Ideally, the modal closes when one clicks on the signup button, and then they are redirected to the signup page.